pyproject.toml
[project]
name = "fastapi-backbone-uv"
version = "0.1.0"
description = "A basic FastAPI backbone for testing router logic."
authors = [
    { name = "Your Name", email = "you@example.com" }
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "fastapi-cli>=0.0.4,<1.0.0",
    "loguru>=0.7.3",
    "msal>=1.32.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
]

[tool.uv]
# This section is for uv-specific configuration if needed
